+++ LessonsAndersen/src/main/java/lesson15/task1/PageLogin/login.java
@@ -1,11 +1,11 @@
-package lesson15.testcases.PageLogin;
+package lesson15.task1.PageLogin;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
-public class testCase1 {
+public class login {
     public static void main(String[] args) throws InterruptedException {
         WebDriverManager.chromedriver().setup();
         WebDriver driver = new ChromeDriver();
@@ -13,18 +13,11 @@ public class testCase1 {
         driver.get("https://qa-course-01.andersenlab.com/login");
         Thread.sleep(1000);
         driver.findElement(By.name("email")).sendKeys("S.Iaffwnfeffdsorv@gmail.com");
-//        driver.findElement(By.id("login-email")).sendKeys("sample@gmail.com");
         Thread.sleep(1000);
 
         driver.findElement(By.name("password")).sendKeys("string123");
         Thread.sleep(1000);
-
-        driver.findElement(By.name("submit")).click();
-        Thread.sleep(1000);
-
-//        driver.findElement(By.id("login-password")).sendKeys("Sample@123");
-//        driver.findElement(By.xpath("//button[text()='Sing   in']")).click();
-//        time.sleep(2000);
-
+        driver.findElement(By.cssSelector("button[type='submit']")).click();
+        driver.quit();
     }
 }
