+++ LessonsAndersen/src/main/java/lesson15/task1/PageLogin/redirectingRegistration.java
@@ -1,17 +1,21 @@
-package lesson15.testcases.PageRegistration;
+package lesson15.task1.PageLogin;
+
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
-public class signInRedirection {
+public class redirectingRegistration {
     public static void main(String[] args) throws InterruptedException {
         WebDriverManager.chromedriver().setup();
         WebDriver driver = new ChromeDriver();
         driver.manage().window().maximize();
-        driver.get("https://qa-course-01.andersenlab.com/registration");
+        driver.get("https://qa-course-01.andersenlab.com/login");
         Thread.sleep(1000);
-        driver.findElement(By.xpath("//span[text()='Sing in']")).click();
+
+        driver.findElement(By.linkText("Registration")).click();
+                Thread.sleep(1000);
+
+                driver.quit();
     }
 }
-
