+++ LessonsAndersen/src/main/java/lesson15/task1/PageRegistration/signInRedirection.java
@@ -1,20 +1,18 @@
-package lesson15.testcases.PageLogin;
-
+package lesson15.task1.PageRegistration;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
-public class redirectingRegistration {
+public class signInRedirection {
     public static void main(String[] args) throws InterruptedException {
         WebDriverManager.chromedriver().setup();
         WebDriver driver = new ChromeDriver();
         driver.manage().window().maximize();
-        driver.get("https://qa-course-01.andersenlab.com/login");
+        driver.get("https://qa-course-01.andersenlab.com/registration");
         Thread.sleep(1000);
-
-        driver.findElement(By.linkText("Registration")).click();
-                Thread.sleep(1000);
-
+        driver.findElement(By.xpath("//span[text()='Sing in']")).click();
+        driver.quit();
     }
 }
+
