--- LessonsAndersen/src/main/java/ru/andersen/courses/java/aza3345/lessonauto/App.java
@@ -1,13 +1,0 @@
-package ru.andersen.courses.java.aza3345.lessonauto;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
